Don't lose previously parsed text & html parts when going through nested multipart/mixed mimeparts
I tried to do the same for multipart/mixed that contains multipart/related,
but because the semantics of those are a bit more complex in regards to content-disposition
(eg. plain-text attachments), this isn't as simple.

We haven't seen them in the wild yet, so I don't want to break stuff without actually fixing a real problem.